Refactor spn package architecture
